# New integration tests framework