--- a/doc/sphinx/zrefs.bib
+++ b/doc/sphinx/zrefs.bib
@@ -503,6 +503,16 @@ school = {Universit\"{a}t Bochum},
 year = {2011}
 }
 
+@Book{kruger12a,
+author={Kr\"{u}ger, Timm},
+title={{Computer simulation study of collective phenomena in dense suspensions of red blood cells under shear}},
+year={2012},
+publisher={Vieweg+Teubner Verlag},
+address={Wiesbaden},
+isbn={978-3-8348-2376-2},
+doi={10.1007/978-3-8348-2376-2},
+}
+
 @Article{landsgesell17a,
 title                      = {{W}ang-{L}andau Reaction Ensemble Method: {S}imulation of Weak Polyelectrolytes and General Acid-Base Reactions},
 author                     = {Landsgesell, Jonas and Holm, Christian and Smiatek, Jens},