--- a/src/core/immersed_boundary/ibm_triel.hpp
+++ b/src/core/immersed_boundary/ibm_triel.hpp
@@ -37,14 +37,6 @@ int IBM_Triel_SetParams(int bond_type, int ind1, int ind2, int ind3,
                         double maxDist, tElasticLaw elasticLaw, double k1,
                         double k2);
 
-/** Update the IBM Triel parameters from a checkpoint.
- *  Ideas:
- *  - parameters are set in the run-continue script
- *  - also reference shape is recomputed there
- *  - only pass two values here to check consistency
- */
-int IBM_Triel_ResetParams(int bond_type, double k1, double l0);
-
 /** Calculate the forces.
  *  The equations can be found in Appendix C of @cite kruger12a.
  *  @return the forces on @p p1, @p p2, @p p3