--- a/src/core/particle_data.cpp
+++ b/src/core/particle_data.cpp
@@ -120,7 +120,7 @@ using UpdatePropertyMessage = boost::variant
                          &Prop::vs_relative>
 #endif
 #endif
-#ifdef LANGEVIN_PER_PARTICLE
+#if defined(LANGEVIN_PER_PARTICLE) || defined(BROWNIAN_PER_PARTICLE)
         , UpdateProperty<double, &Prop::T>
 #ifndef PARTICLE_ANISOTROPY
         , UpdateProperty<double, &Prop::gamma>
@@ -134,7 +134,7 @@ using UpdatePropertyMessage = boost::variant
         , UpdateProperty<Utils::Vector3d, &Prop::gamma_rot>
 #endif // PARTICLE_ANISOTROPY
 #endif // ROTATION
-#endif // LANGEVIN_PER_PARTICLE
+#endif // LANGEVIN_PER_PARTICLE || BROWNIAN_PER_PARTICLE
 #ifdef EXTERNAL_FORCES
         , UpdateProperty<uint8_t, &Prop::ext_flag>
         , UpdateProperty<Utils::Vector3d, &Prop::ext_force>
@@ -864,7 +864,7 @@ void set_particle_torque_lab(int part, const Utils::Vector3d &torque_lab) {
 }
 #endif
 
-#ifdef LANGEVIN_PER_PARTICLE
+#if defined(LANGEVIN_PER_PARTICLE) || defined(BROWNIAN_PER_PARTICLE)
 void set_particle_temperature(int part, double T) {
   mpi_update_particle_property<double, &ParticleProperties::T>(part, T);
 }
@@ -893,7 +893,7 @@ void set_particle_gamma_rot(int part, Utils::Vector3d gamma_rot) {
 }
 #endif // PARTICLE_ANISOTROPY
 #endif // ROTATION
-#endif // LANGEVIN_PER_PARTICLE
+#endif // LANGEVIN_PER_PARTICLE || BROWNIAN_PER_PARTICLE
 
 #ifdef EXTERNAL_FORCES
 #ifdef ROTATION
@@ -1242,7 +1242,7 @@ void pointer_to_fix(Particle const *p, const uint8_t *&res) {
 }
 #endif
 
-#ifdef LANGEVIN_PER_PARTICLE
+#if defined(LANGEVIN_PER_PARTICLE) || defined(BROWNIAN_PER_PARTICLE)
 void pointer_to_gamma(Particle const *p, double const *&res) {
 #ifndef PARTICLE_ANISOTROPY
   res = &(p->p.gamma);
@@ -1264,7 +1264,7 @@ void pointer_to_gamma_rot(Particle const *p, double const *&res) {
 void pointer_to_temperature(Particle const *p, double const *&res) {
   res = &(p->p.T);
 }
-#endif // LANGEVIN_PER_PARTICLE
+#endif // LANGEVIN_PER_PARTICLE || BROWNIAN_PER_PARTICLE
 
 #ifdef ENGINE
 void pointer_to_swimming(Particle const *p,