--- a/src/core/particle_data.hpp
+++ b/src/core/particle_data.hpp
@@ -240,7 +240,7 @@ void set_particle_vs_relative(int part, int vs_relative_to, double vs_distance,
                               Utils::Vector4d const &rel_ori);
 #endif
 
-#ifdef LANGEVIN_PER_PARTICLE
+#if defined(LANGEVIN_PER_PARTICLE) || defined(BROWNIAN_PER_PARTICLE)
 /** Call only on the master node: set particle temperature.
  *  @param part the particle.
  *  @param T its new temperature.
@@ -263,7 +263,7 @@ void set_particle_gamma_rot(int part, double gamma);
 void set_particle_gamma_rot(int part, Utils::Vector3d gamma_rot);
 #endif
 #endif
-#endif // LANGEVIN_PER_PARTICLE
+#endif // LANGEVIN_PER_PARTICLE || BROWNIAN_PER_PARTICLE
 
 #ifdef EXTERNAL_FORCES
 #ifdef ROTATION
@@ -389,15 +389,13 @@ void pointer_to_ext_torque(Particle const *p, double const *&res2);
 void pointer_to_fix(Particle const *p, const uint8_t *&res);
 #endif
 
-#ifdef LANGEVIN_PER_PARTICLE
+#if defined(LANGEVIN_PER_PARTICLE) || defined(BROWNIAN_PER_PARTICLE)
 void pointer_to_gamma(Particle const *p, double const *&res);
 void pointer_to_temperature(Particle const *p, double const *&res);
 #ifdef ROTATION
 void pointer_to_gamma_rot(Particle const *p, double const *&res);
 #endif
-#endif // LANGEVIN_PER_PARTICLE
-#ifdef ROTATION
-#endif
+#endif // LANGEVIN_PER_PARTICLE || BROWNIAN_PER_PARTICLE
 
 #ifdef ENGINE
 void pointer_to_swimming(Particle const *p,