--- a/src/core/stokesian_dynamics/sd_interface.hpp
+++ b/src/core/stokesian_dynamics/sd_interface.hpp
@@ -34,13 +34,13 @@
 
 #include <unordered_map>
 
-void set_sd_viscosity(double eta);
+int set_sd_viscosity(double eta);
 double get_sd_viscosity();
 
-void set_sd_radius_dict(std::unordered_map<int, double> const &x);
+int set_sd_radius_dict(std::unordered_map<int, double> const &x);
 std::unordered_map<int, double> get_sd_radius_dict();
 
-void set_sd_kT(double kT);
+int set_sd_kT(double kT);
 double get_sd_kT();
 
 void set_sd_flags(int flg);