--- a/src/core/stokesian_dynamics/sd_interface.hpp
+++ b/src/core/stokesian_dynamics/sd_interface.hpp
@@ -34,13 +34,13 @@
 
 #include <unordered_map>
 
-int set_sd_viscosity(double eta);
+void set_sd_viscosity(double eta);
 double get_sd_viscosity();
 
-int set_sd_radius_dict(std::unordered_map<int, double> const &x);
+void set_sd_radius_dict(std::unordered_map<int, double> const &x);
 std::unordered_map<int, double> get_sd_radius_dict();
 
-int set_sd_kT(double kT);
+void set_sd_kT(double kT);
 double get_sd_kT();
 
 void set_sd_flags(int flg);