--- a/src/python/espressomd/particle_data.pxd
+++ b/src/python/espressomd/particle_data.pxd
@@ -129,7 +129,7 @@ cdef extern from "particle_data.hpp":
         void set_particle_virtual(int part, int isVirtual)
         void pointer_to_virtual(const particle * P, const bint * & res)
 
-    IF LANGEVIN_PER_PARTICLE:
+    IF LANGEVIN_PER_PARTICLE or BROWNIAN_PER_PARTICLE:
         void set_particle_temperature(int part, double T)
         void pointer_to_temperature(const particle * p, const double * & res)