Fixes #3883

Description of changes:
- write more thorough RNG checks for thermostats
- skip slow/statistical tests in coverage and sanitizer builds
- fix broken ifdefs for Brownian Dynamics
- cleanup IBM code