@jngrad 2020-11-19
Here is the summary for the python testsuite, using the same CI runner when possible:

| CI job           | Old runtime | New runtime | speed-up  |
| ---------------- |-------------| ----------- | --------- |
| clang-sanitizer  | 2109.2s     | 988.1s      | 19m (53%) |
| no_rotation      | 647.7s      | 468.5s      | 3m (28%)  |
| maxset           | 961.6s      | 616.0s      | 6m (36%)  |
| default          | 970.5s      | 771.3s      | 3m (20%)  |
| cuda11-maxset    | 1113.4s     | 774.9s      | 6m (30%)  |

The complete pipeline runtime went from 72 min to 38 min. The Clang job completes in 35 min instead of 50 min.

@RudolfWeeber your plan in https://github.com/espressomd/espresso/issues/3883#issuecomment-716630623 was really good, however I had to deviate from it slightly:
* `virtual_sites_tracers` is needed to get code coverage for IBM particle coupling to the fluid (not sure why though, the LB actor is removed before the IBM bonds are created)
* Brownian/NpT were moved fully, but new files with minimum checks were introduced for both integrators, otherwise their code coverage would be 0%

Other remarks:
* `make check_python` still works the same way as before
* `make check_python_skip_long` to run the subset of fast tests
* the comment lines in the thermostat tests about RNG seed and counter state were obtained using [cout.diff](https://github.com/espressomd/espresso/files/5568591/cout.diff.txt)