@jngrad 2020-11-20
The slowest tests are now:
* `stokesian_dynamics`: 1 min 10 sec in sanitizer, 40 sec in coverage, 20-30s in others
   * now skipped in sanitizer/coverage builds
   * new test added to check the SD interface
* `elc_vs_analytic`: 1 min 20 sec in sanitizer/coverage
   * removing it from coverage builds would require writing a small test without constant potential to preserve code coverage
* reaction ensemble tests:
   * `constant_pH`: 1 min 30 sec in sanitizer, 45 sec in coverage
   * `reaction_ensemble`: 30-40 sec in sanitizer/coverage, 5 sec in others
   * `widom_insertion`: 30-40 sec in sanitizer/coverage, 5 sec in others
   * `wang_landau_reaction_ensemble`: already skipped in sanitizer/coverage
   * long-term plan would be to skip these tests in sanitizer/coverage and write much smaller tests to check the RE interface (error messages, output files, and run a few reaction steps), like I did here for the thermostats and integrators